src/dash/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
boto3==1.37.12
botocore==1.37.12
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
folium==0.19.5
future==1.0.0
geographiclib==2.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.30.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
ratelim==0.1.6
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.11.4
six==1.17.0
smmap==5.0.2
streamlit==1.43.2
streamlit-autorefresh==1.0.1
streamlit-cookies-manager==0.2.0
streamlit_folium==0.24.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
xyzservices==2025.1.0
